move navbar to shared folder
